requirements.txt
Flask
gunicorn
pandas
numpy
matplotlib
sklearn
ipykernel
pyYAML
requests
dill
evidently
urllib3
-e .

